Android-Quiz-App
This project is built using java and android libraries where the user
can upload set of questions and answers and ask his peers to take up the
test.
At the end of each test execution,results will be displayed.
This mini-project can be used as an internal assessment tool,a quiz
game.